requirements.txt
seaborn>=0.9
matplotlib-base>=3.1.3
pandas>=1
scikit-learn>=0.22
scipy>=1.4
adjustText>=0.6.0
